Conflicts:
	src/Arch/Arm/AArch64.cs
	src/Arch/Arm/Arm32ProcessorArchitecture.cs
	src/Arch/Arm/ThumbProcessorArchitecture.cs
	src/Arch/M68k/M68kArchitecture.cs
	src/Arch/Mips/MipsProcessorArchitecture.cs
	src/Arch/Mos6502/Mos6502ProcessorArchitecture.cs
	src/Arch/Pdp11/Pdp11Architecture.cs
	src/Arch/PowerPC/PowerPcArchitecture.cs
	src/Arch/Sparc/SparcArchitecture.cs
	src/Arch/Vax/VaxArchitecture.cs
	src/Arch/X86/ProcessorMode.cs
	src/Arch/X86/X86ProcessorArchitecture.cs
	src/Arch/X86/X86Rewriter.Alu.cs
	src/Arch/X86/X86Rewriter.Fpu.cs
	src/Arch/Xtensa/XtensaArchitecture.cs
	src/Arch/Z80/Z80ProcessorArchitecture.cs
	src/Core/CodeEmitter.cs
	src/Core/Core.csproj
	src/Core/IProcessorArchitecture.cs
	src/Core/IStorageBinder.cs
	src/Decompiler/Analysis/GlobalCallRewriter.cs
	src/UnitTests/Analysis/CallRewriterTests.cs
	src/UnitTests/Analysis/GlobalCallRewriterTests.cs
	src/UnitTests/Analysis/LongAddRewriterTests.cs
	src/UnitTests/Analysis/SsaTransformTests.cs
	src/UnitTests/Analysis/ValuePropagationTests.cs
	src/UnitTests/Core/ProcedureStateTests.cs
	src/UnitTests/Evaluation/ConstDivisionImplementedByMultiplicationTests.cs
	src/UnitTests/Loading/LoaderTests.cs
	src/UnitTests/Mocks/ArrayMock.cs
	src/UnitTests/Mocks/FakeArchitecture.cs
	src/UnitTests/Mocks/ProcedureBuilder.cs
	src/UnitTests/Scanning/BlockWorkitem.X86Tests.cs
	src/tests/Intel/RwDivideTests.exp
	src/tests/Structure/CccComplexText.exp
	src/tools/c2xml/UnitTests/XmlConverterTests.cs
	subjects/PE-x86-64/varargs_test/varargs_test.globals.c
	subjects/PE-x86-64/varargs_test/varargs_test.h
	subjects/regressions/reko-90/PP.c
	subjects/regressions/reko-90/PP.h

+++ src/Decompiler/Analysis/LongAddRewriter.cs
@@ -563,7 +563,7 @@ namespace Reko.Analysis
                 sidDstLo.DefStatement = stmCastLo;
 
                 var sidDstHi = GetSsaIdentifierOf(hiCandidate.Dst);
-                var slice = new Slice(hiCandidate.Dst.DataType, dst, (uint)loCandidate.Dst.DataType.BitSize);
+                var slice = new Slice(hiCandidate.Dst.DataType, dst, loCandidate.Dst.DataType.BitSize);
                 var stmSliceHi = stmts.Insert(iStm++, linAddr, new Assignment(
                     sidDstHi.Identifier, slice));
                 var stmDeadHi = sidDstHi.DefStatement;
